refactor: migrate to Next.js API Routes
