--- api/.eslintrc
@@ -1,11 +1,0 @@
-{
-  "extends": "../.eslintrc.js",
-  "overrides": [
-    {
-      "files": ["**/*.ts", "**/*.tsx"],
-      "parserOptions": {
-        "project": "./tsconfig.json"
-      }
-    }
-  ]
-}
