--- api/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@pinjollist/api",
-  "version": "1.0.0",
-  "description": "The Pinjollist API endpoint.",
-  "main": "N/A",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/resir014/pinjollist.git"
-  },
-  "author": "Resi Respati <resir014@gmail.com>",
-  "license": "Apache-2.0",
-  "private": true,
-  "engines": {
-    "node": "10.x"
-  },
-  "scripts": {
-    "lint": "eslint \"**/*.{js,ts}\"",
-    "type-check": "tsc --noEmit",
-    "validate": "yarn lint && yarn type-check"
-  },
-  "dependencies": {
-    "google-spreadsheet": "^2.0.8",
-    "micro": "^9.3.3"
-  },
-  "devDependencies": {
-    "@types/micro": "^7.3.3"
-  }
-}
