+++ package.json
@@ -11,7 +11,6 @@
   "license": "Apache-2.0",
   "private": true,
   "workspaces": [
-    "api",
     "scripts",
     "www"
   ],
