+++ www/package.json
@@ -17,7 +17,9 @@
     "date-fns": "^2.2.1",
     "debounce-fn": "^3.0.0",
     "fuse.js": "^3.4.2",
+    "google-spreadsheet": "^2.0.8",
     "isomorphic-unfetch": "^3.0.0",
+    "micro": "^9.3.4",
     "next": "^9.1.2",
     "next-compose-plugins": "^2.1.1",
     "normalize.css": "^8.0.1",
@@ -29,6 +31,7 @@
   },
   "devDependencies": {
     "@types/debounce-fn": "^3.0.0",
+    "@types/micro": "^7.3.3",
     "@types/nprogress": "^0.2.0",
     "@types/react": "16.9.11",
     "@types/react-dom": "16.9.3",
