+++ www/types/common.ts
@@ -1,11 +1,11 @@
-export interface APIResponse<T = {}> {
-  status: 'ok';
+export type ResponseStatuses = 'ok' | 'error';
+
+export interface ResponseObject<T = {}> {
+  status: ResponseStatuses;
   data: T;
 }
 
-export interface ErrorAPIResponse {
-  status: 'error';
-  data: {
-    message: string;
-  };
+export interface ServiceAccount {
+  client_email: string;
+  private_key: string;
 }
